Changing a password upgrades from old hash to new hash. Old Hash: $tmp = strtolower($pwd); $oldHash = md5($tmp); New Hash: $tmp = strtolower($pwd.$uid); $newHash = sha1(sha1($tmp).$tmp);
There is code that is currently commented out that will update a user
from the old hash to the new hash when they log in.  Wait to hear
if there are problems when changing passwords before implementing this
code.  This code will be enabled in a future version.

Seems a benign change but no harm in doing this in stages.

Both 32 character and 40 character hash will work so this code is backward
compatible to old style hash and gradually moves users to the new style
hash.

 Modified Files:
 	churchinfo/Default.php churchinfo/UserPasswordChange.php
-Ed